Use docker for dev environment
1) Rename currently-hardcoded hostname/ports
2) Default dockercontainer builds an uberjar, but we can spawn a REPL
for dev
3) Decided, after much research and little findings, that if we want a
"reloaded.repl" experience for dev, we can't put REPLs into the system
but have to use plain "lein repl". This is because we can't have the
REPL stop/start if we're using it presumably. I don't know if this is
in fact the case or not, but it seems intuitively so. Embedded REPLs
seem to be more for production environments, which is reasonable?
4) If we're running the REPL in a dockerized process, I have to
explicitly include any nrepl middleware dependencies I need (refactor,
piggieback, cider)
5) I am not entirely convinced figwheel is properly picking up file
changes.
6) I'm not entirely sure what the benefits of doing this are, in all
honesty. I have to bring up docker-compose for postgres anyway, and I
still have to connect in via Emacs but now I have to type more and
things work less seamlessly. But I'll try it out and decide if it's
worth it.